requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
alpha-vantage>=2.3.1
finnhub-python>=2.4.0
PyYAML>=6.0